# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "POps central server API. `/api/v1` is the stable surface for integrations; the same endpoints answer under plain `/api` for the panel and existing agents. Authenticate with an API token (`Authorization: Bearer pops_...`, created by a superadmin in Ayarlar → Güvenlik) or a panel login JWT. Paths marked deprecated still work; use their REST names. Reference: docs/api.md.",
    "title": "POps API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/activity/agent/{pc_name}": {
      "get": {
        "operationId": "agent_activity_api_activity_agent__pc_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "pc_name",
            "required": true,
            "schema": {
              "title": "Pc Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Activity",
        "tags": [
          "activity"
        ]
      }
    },
...
  }
}